# Make prevector::resize() and other prevector operations much faster