While profiling the `loadblk` thread I noticed that a lot of time was being spent in `prevector::resize()`. I have some data here indicating that it takes up **37%** of the time in `ReadBlockFromDisk()`: https://monad.io/readblockfromdisk.svg

This branch improves things significantly. According to the prevector benchmark, the speedups are:
 * `prevector::clear()` for trivial types becomes 8.1x faster
 * `prevector::~prevector()` for trivial types becomes 7.6x faster (most of its time was spent in `resize()`)
 * `prevector::resize()` for trivial types becomes 17.9x faster
 * `prevector::resize()` for non-trivial types becomes 2.2x faster

Note that in practice it looks like the prevector is only used to contain `unsigned char` types, which is a trivial type. The benchmarks are testing a bit of an extreme case, but the changes here are motivated by the profiling data for `ReadBlockFromDisk()` I linked to above.

The pull request here is two commits, the first commit adds the new benchmarks, and the second commit actually optimizes the prevector implementation. The numbers given above are based on me comparing the benchmark numbers from the first commit (which is basically benchmarking master) against the benchmark numbers in the second commit:

Before prevector optimizations:
```
$ ./src/bench/bench_bitcoin -filter='Prevector.*'
# Benchmark, evals, iterations, total, min, max, median
PrevectorClear, 5, 5600, 3.51216, 0.000124847, 0.000126195, 0.000125358
PrevectorDestructor, 5, 5700, 3.32806, 0.000115869, 0.000117395, 0.00011691
PrevectorResizeNontrivial, 5, 4100, 3.41065, 0.000162157, 0.000171089, 0.0001671
PrevectorResizeTrivial, 5, 4900, 3.51391, 0.000137509, 0.000150316, 0.00014522
```

After my optimizations:
```
$ ./src/bench/bench_bitcoin -filter='Prevector.*'
# Benchmark, evals, iterations, total, min, max, median
PrevectorClear, 5, 45000, 3.50816, 1.52231e-05, 1.63413e-05, 1.54451e-05
PrevectorDestructor, 5, 42800, 3.29708, 1.53111e-05, 1.54935e-05, 1.54058e-05
PrevectorResizeNontrivial, 5, 8700, 3.32136, 7.61098e-05, 7.66597e-05, 7.6256e-05
PrevectorResizeTrivial, 5, 84000, 3.45793, 7.96525e-06, 8.79618e-06, 8.11656e-06
```

I also updated the iteration count in the benchmarks a bit, to make them take about the same time as they currently do.