@eklitzke 2018-02-27
It looks like this branch is failing because `std::is_trivially_constructible` wasn't added until GCC 5 (according to https://www.gnu.org/software/gcc/gcc-5/changes.html ), even though it is part of C++11. I can find a workaround for GCC 4.8.

This isn't quite the same as #11988 because that still uses placement new.