@eklitzke 2018-02-27
This new branch cherry-picks @AkioNak 's commit from #11988 and then adds my memset optimization. I constructed this PR such that it has three commits:
 * First commit just adds new benchmarks
 * Second commit is from #11988 (unmodified, preserves authorship)
 * Third commit implements my `memset()` optimization

I have a script that benchmarks the performance for each of the three commits. Here's the summarized data. The first commit is considered to have a baseline speedup of 1x, and then each of the subsequent commits shows the speedup from the base:

```
34d238     PrevectorClear                 1.00x
34d238     PrevectorDestructor            1.00x
34d238     PrevectorResizeNontrivial      1.00x
34d238     PrevectorResizeTrivial         1.00x
d1d28e     PrevectorClear                 2.36x
d1d28e     PrevectorDestructor            2.16x
d1d28e     PrevectorResizeNontrivial      3.24x
d1d28e     PrevectorResizeTrivial         2.76x
f8b078     PrevectorClear                 8.31x
f8b078     PrevectorDestructor            7.62x
f8b078     PrevectorResizeNontrivial      6.56x
f8b078     PrevectorResizeTrivial         18.99x
```

Here's the raw data I generated the above from:
```
git 34d2387b1d5c753d6ba130eb6b4730dcc15072c5
# Benchmark, evals, iterations, total, min, max, median
PrevectorClear, 5, 5600, 3.47561, 0.00012317, 0.000125752, 0.000123494
PrevectorDestructor, 5, 5700, 3.2223, 0.000112914, 0.000113208, 0.000113068
PrevectorResizeNontrivial, 5, 4100, 3.24072, 0.000158005, 0.000158212, 0.000158075
PrevectorResizeTrivial, 5, 4900, 3.24407, 0.000131094, 0.000133749, 0.000132495

git d1d28e72154a9f2e7fb56001f2b790074d1cfa30
# Benchmark, evals, iterations, total, min, max, median
PrevectorClear, 5, 5600, 1.4682, 5.23405e-05, 5.2611e-05, 5.24265e-05
PrevectorDestructor, 5, 5700, 1.49587, 5.24034e-05, 5.26179e-05, 5.2448e-05
PrevectorResizeNontrivial, 5, 4100, 1.00747, 4.87854e-05, 5.04511e-05, 4.88389e-05
PrevectorResizeTrivial, 5, 4900, 1.17581, 4.79055e-05, 4.81338e-05, 4.79753e-05

git f8b078c8cda6d4ac6e3b112ce605b08f6ac002b9
# Benchmark, evals, iterations, total, min, max, median
PrevectorClear, 5, 5600, 0.419837, 1.48182e-05, 1.55947e-05, 1.48556e-05
PrevectorDestructor, 5, 5700, 0.422994, 1.48187e-05, 1.48808e-05, 1.48327e-05
PrevectorResizeNontrivial, 5, 4100, 0.494272, 2.40686e-05, 2.41813e-05, 2.4084e-05
PrevectorResizeTrivial, 5, 4900, 0.171037, 6.89859e-06, 7.06054e-06, 6.97555e-06
```