@laanwj 2018-02-27
Ping @sipa who wrote this code.