@eklitzke 2018-02-27
Good catch, I only tested the last commit. I'll rebase this branch so all of the commits build cleanly, and move the macro definition.