@eklitzke 2018-02-27
Here are the latest results, tested under both GCC 7.3 and GCC 4.8. For both versions I ran the full test suite and benchmarks.

### GCC 4.8

```
Commit     Test                                     Median (s)   Speedup
------------------------------------------------------------------------
f0e7aa     PrevectorClearNontrivial                 8.15034e-05   1.00x
f0e7aa     PrevectorClearTrivial                    1.49740e-04   1.00x
f0e7aa     PrevectorDestructorNontrivial            8.27660e-05   1.00x
f0e7aa     PrevectorDestructorTrivial               1.48346e-04   1.00x
f0e7aa     PrevectorResizeNontrivial                8.26464e-05   1.00x
f0e7aa     PrevectorResizeTrivial                   1.50579e-04   1.00x
e46be2     PrevectorClearNontrivial                 2.59100e-05   3.15x
e46be2     PrevectorClearTrivial                    4.80695e-05   3.12x
e46be2     PrevectorDestructorNontrivial            2.54231e-05   3.26x
e46be2     PrevectorDestructorTrivial               4.80365e-05   3.09x
e46be2     PrevectorResizeNontrivial                2.54693e-05   3.24x
e46be2     PrevectorResizeTrivial                   4.82063e-05   3.12x
3f9abf     PrevectorClearNontrivial                 2.60759e-05   3.13x
3f9abf     PrevectorClearTrivial                    6.08490e-06  24.61x
3f9abf     PrevectorDestructorNontrivial            2.55872e-05   3.23x
3f9abf     PrevectorDestructorTrivial               6.09886e-06  24.32x
3f9abf     PrevectorResizeNontrivial                2.56024e-05   3.23x
3f9abf     PrevectorResizeTrivial                   6.15214e-06  24.48x
```

### GCC 7.3
```
Commit     Test                                     Median (s)   Speedup
------------------------------------------------------------------------
f0e7aa     PrevectorClearNontrivial                 1.58989e-04   1.00x
f0e7aa     PrevectorClearTrivial                    1.31495e-04   1.00x
f0e7aa     PrevectorDestructorNontrivial            1.58322e-04   1.00x
f0e7aa     PrevectorDestructorTrivial               1.31456e-04   1.00x
f0e7aa     PrevectorResizeNontrivial                1.58518e-04   1.00x
f0e7aa     PrevectorResizeTrivial                   1.29851e-04   1.00x
e46be2     PrevectorClearNontrivial                 4.89286e-05   3.25x
e46be2     PrevectorClearTrivial                    4.81217e-05   2.73x
e46be2     PrevectorDestructorNontrivial            4.94412e-05   3.20x
e46be2     PrevectorDestructorTrivial               4.79828e-05   2.74x
e46be2     PrevectorResizeNontrivial                4.89689e-05   3.24x
e46be2     PrevectorResizeTrivial                   4.80065e-05   2.70x
3f9abf     PrevectorClearNontrivial                 2.59529e-05   6.13x
3f9abf     PrevectorClearTrivial                    6.39799e-06  20.55x
3f9abf     PrevectorDestructorNontrivial            2.58583e-05   6.12x
3f9abf     PrevectorDestructorTrivial               6.40997e-06  20.51x
3f9abf     PrevectorResizeNontrivial                2.56362e-05   6.18x
3f9abf     PrevectorResizeTrivial                   6.40775e-06  20.26x
```

It's interesting that GCC 4.8 generates better code in some cases. Anyway, hopefully this is enough data to convince people that this is a good change.